requirements.txt
fastapi>=0.110
uvicorn>=0.29
python-multipart>=0.0.9
torch>=2.1.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.0
Pillow>=10.0.0
opencv-python-headless<4.12
numpy>=1.24,<2
mlflow>=2.9.0
pytest>=7.4.0
alibi-detect[tensorflow]>=0.13
onnx>=1.16
onnxruntime>=1.18
onnxscript>=0.1
